# swagger/swagger.yaml
definitions:
  models.MetricType:
    enum:
    - gauge
    - counter
    type: string
    x-enum-varnames:
    - Gauge
    - Counter
  v2handlers.GetMetricsRequest:
    properties:
      metricID:
        description: Идентификатор метрики
        example: Alloc
        type: string
      metricType:
        allOf:
        - $ref: '#/definitions/models.MetricType'
        description: Тип метрики
        example: gauge
    type: object
  v2handlers.Metrics:
    properties:
      delta:
        description: значение метрики в случае передачи counter
        example: 5
        type: integer
      id:
        description: идентификатор метрики
        example: Alloc
        type: string
      type:
        description: параметр, принимающий значение gauge или counter
        example: gauge
        type: string
      value:
        description: значение метрики в случае передачи gauge
        example: 12.6
        type: number
    type: object
info:
  contact: {}
paths:
...
  /value:
    post:
      consumes:
      - application/json
      description: Получение метрики по ID
      operationId: getMetric
      parameters:
      - description: Тело запроса
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v2handlers.GetMetricsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2handlers.Metrics'
        "400":
          description: Неверный запрос
          schema:
            type: string
        "404":
          description: Метрика не найдена
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      summary: Получение метрики
      tags:
      - Metrics
...
swagger: "2.0"